+++ Scripts/Models/Common/Collaborations.cs
@@ -4,7 +4,7 @@ public class Collaborations : BaseEntity
     public string Name { get; set; }
     public string Image { get; set; }
     public string Rare { get; set; }
-    public int Quality { get; set; }
+    public double Quality { get; set; }
     public int CurrentStar { get; set; }
     public int TempStar { get; set; }
     public int Star { get; set; }
