+++ Scripts/Models/Common/Skills.cs
@@ -6,7 +6,7 @@ public class Skills : BaseEntity
     public string Name { get; set; }
     public string Image { get; set; }
     public string Rare { get; set; }
-    public int Quality { get; set; }
+    public double Quality { get; set; }
     public string Type { get; set; }
     public int CurrentStar { get; set; }
     public int TempStar { get; set; }
