+++ Scripts/Repositories/Gallery/BadgesGalleryRepository.cs
@@ -50,7 +50,7 @@ public class BadgesGalleryRepository : IBadgesGalleryRepository
                                 Name = reader.GetString("name"),
                                 Image = reader.GetString("image"),
                                 Rare = reader.GetString("rare"),
-                                Quality = reader.GetInt32("quality"),
+                                Quality = reader.GetDouble("quality"),
                                 Power = reader.GetDouble("power"),
                                 Health = reader.GetDouble("health"),
                                 PhysicalAttack = reader.GetDouble("physical_attack"),
@@ -91,7 +91,7 @@ public class BadgesGalleryRepository : IBadgesGalleryRepository
                                 IgnoreReflectionRate = reader.GetDouble("ignore_reflection_rate"),
                                 ReflectionDamageRate = reader.GetDouble("reflection_damage_rate"),
                                 ReflectionResistanceRate = reader.GetDouble("reflection_resistance_rate"),
-                                Mana = reader.GetFloat("mana"),
+                                Mana = reader.GetDouble("mana"),
                                 ManaRegenerationRate = reader.GetDouble("mana_regeneration_rate"),
                                 DamageToDifferentFactionRate = reader.GetDouble("damage_to_different_faction_rate"),
                                 ResistanceToDifferentFactionRate = reader.GetDouble("resistance_to_different_faction_rate"),
@@ -656,7 +656,7 @@ public class BadgesGalleryRepository : IBadgesGalleryRepository
                     {
                         sumBadges.Power = reader["total_power"] as double? ?? 0;
                         sumBadges.Health = reader["total_health"] as double? ?? 0;
-                        sumBadges.Mana = reader["total_mana"] as float? ?? 0f;
+                        sumBadges.Mana = reader["total_mana"] as double? ?? 0f;
 
                         sumBadges.PhysicalAttack = reader["total_physical_attack"] as double? ?? 0;
                         sumBadges.PhysicalDefense = reader["total_physical_defense"] as double? ?? 0;
